Update NuGet package versioning format
Updated the `dotnet pack` command in the release pipeline to use `v${{ steps.gitversion.outputs.MajorMinorPatch }}` for the `PackageVersion` property instead of `${{ steps.gitversion.outputs.NuGetVersionV3 }}`. This change simplifies the versioning scheme.

Retained the conditional check in the `Publish to NuGet.org` step to skip publishing if the `NUGET_API_KEY` secret is not set.